+++ src/Video/VideoThumbGeneratorInterface.php
@@ -9,7 +9,7 @@ use Symfony\Component\Process\Process;
 
 interface VideoThumbGeneratorInterface
 {
-    public function getSymfonyProcess(string $videoFile, string $thumbnailFile, ThumbParamsInterface $thumbParams, ?ProcessParamsInterface $processParam = null): Process;
+    public function getSymfonyProcess(string $videoFile, string $thumbnailFile, VideoThumbParamsInterface $thumbParams, ?ProcessParamsInterface $processParam = null): Process;
 
-    public function makeThumbnail(string $videoFile, string $thumbnailFile, ThumbParamsInterface $thumbParams, ?callable $callback = null, ?ProcessParamsInterface $processParam = null): void;
+    public function makeThumbnail(string $videoFile, string $thumbnailFile, VideoThumbParamsInterface $thumbParams, ?callable $callback = null, ?ProcessParamsInterface $processParam = null): void;
 }