+++ tests/unit/Video/ThumbParamsTest.php
@@ -7,8 +7,8 @@ namespace MediaToolsTest\Video;
 use PHPUnit\Framework\TestCase;
 use Soluble\MediaTools\Video\Exception\UnsetParamException;
 use Soluble\MediaTools\Video\SeekTime;
-use Soluble\MediaTools\Video\ThumbParams;
-use Soluble\MediaTools\Video\ThumbParamsInterface;
+use Soluble\MediaTools\Video\VideoThumbParams;
+use Soluble\MediaTools\Video\VideoThumbParamsInterface;
 
 class ThumbParamsTest extends TestCase
 {
@@ -18,7 +18,7 @@ class ThumbParamsTest extends TestCase
 
     public function testMustBeImmutable(): void
     {
-        $params = new ThumbParams();
+        $params = new VideoThumbParams();
         self::assertCount(0, $params->toArray());
         $newParams = $params->withQualityScale(2);
         self::assertCount(0, $params->toArray());
@@ -27,52 +27,52 @@ class ThumbParamsTest extends TestCase
 
     public function testOverwriteParams(): void
     {
-        $params = (new ThumbParams())
+        $params = (new VideoThumbParams())
             ->withNoOverwrite();
 
-        self::assertTrue($params->hasParam(ThumbParams::PARAM_OVERWRITE));
-        self::assertFalse($params->getParam(ThumbParams::PARAM_OVERWRITE));
+        self::assertTrue($params->hasParam(VideoThumbParams::PARAM_OVERWRITE));
+        self::assertFalse($params->getParam(VideoThumbParams::PARAM_OVERWRITE));
 
         $params = $params->withOverwrite();
 
-        self::assertTrue($params->hasParam(ThumbParams::PARAM_OVERWRITE));
-        self::assertTrue($params->getParam(ThumbParams::PARAM_OVERWRITE));
+        self::assertTrue($params->hasParam(VideoThumbParams::PARAM_OVERWRITE));
+        self::assertTrue($params->getParam(VideoThumbParams::PARAM_OVERWRITE));
     }
 
     public function testWithoutParam(): void
     {
-        $params = (new ThumbParams())
+        $params = (new VideoThumbParams())
             ->withQualityScale(1)
             ->withSeekTime(new SeekTime(10.3));
 
-        $newParams = $params->withoutParam(ThumbParams::PARAM_QUALITY_SCALE);
-        self::assertTrue($newParams->hasParam(ThumbParams::PARAM_SEEK_TIME));
-        self::assertFalse($newParams->hasParam(ThumbParams::PARAM_QUALITY_SCALE));
-        self::assertTrue($params->hasParam(ThumbParams::PARAM_QUALITY_SCALE));
+        $newParams = $params->withoutParam(VideoThumbParams::PARAM_QUALITY_SCALE);
+        self::assertTrue($newParams->hasParam(VideoThumbParams::PARAM_SEEK_TIME));
+        self::assertFalse($newParams->hasParam(VideoThumbParams::PARAM_QUALITY_SCALE));
+        self::assertTrue($params->hasParam(VideoThumbParams::PARAM_QUALITY_SCALE));
     }
 
     public function testHasParam(): void
     {
-        $params = (new ThumbParams())
+        $params = (new VideoThumbParams())
             ->withOutputFormat('png')
             ->withQualityScale(1);
 
-        self::assertTrue($params->hasParam(ThumbParams::PARAM_QUALITY_SCALE));
-        self::assertFalse($params->hasParam(ThumbParams::PARAM_SEEK_TIME));
-        self::assertSame('png', $params->getParam(ThumbParams::PARAM_OUTPUT_FORMAT));
+        self::assertTrue($params->hasParam(VideoThumbParams::PARAM_QUALITY_SCALE));
+        self::assertFalse($params->hasParam(VideoThumbParams::PARAM_SEEK_TIME));
+        self::assertSame('png', $params->getParam(VideoThumbParams::PARAM_OUTPUT_FORMAT));
     }
 
     public function testWithTime(): void
     {
-        $params = (new ThumbParams())
+        $params = (new VideoThumbParams())
             ->withTime(1.423);
 
-        self::assertTrue($params->hasParam(ThumbParams::PARAM_SEEK_TIME));
+        self::assertTrue($params->hasParam(VideoThumbParams::PARAM_SEEK_TIME));
 
         /**
          * @var SeekTime
          */
-        $time = $params->getParam(ThumbParams::PARAM_SEEK_TIME);
+        $time = $params->getParam(VideoThumbParams::PARAM_SEEK_TIME);
         self::assertEquals(1.423, $time->getTime());
     }
 
@@ -80,28 +80,28 @@ class ThumbParamsTest extends TestCase
     {
         self::expectException(UnsetParamException::class);
 
-        $params = (new ThumbParams())->withTime(10);
+        $params = (new VideoThumbParams())->withTime(10);
 
-        $params->getParam(ThumbParamsInterface::PARAM_QUALITY_SCALE);
+        $params->getParam(VideoThumbParamsInterface::PARAM_QUALITY_SCALE);
     }
 
     public function testWithBuiltInParam(): void
     {
-        $params = (new ThumbParams())
-            ->withBuiltInParam(ThumbParamsInterface::PARAM_QUALITY_SCALE, 5);
+        $params = (new VideoThumbParams())
+            ->withBuiltInParam(VideoThumbParamsInterface::PARAM_QUALITY_SCALE, 5);
 
         self::assertEquals([
-            ThumbParamsInterface::PARAM_QUALITY_SCALE      => 5,
+            VideoThumbParamsInterface::PARAM_QUALITY_SCALE      => 5,
         ], $params->toArray());
     }
 
     public function testWithParamsMustBeIdenticalToConstrutorInject(): void
     {
-        $injectedParams = new ThumbParams([
-            ThumbParams::PARAM_QUALITY_SCALE => 1,
+        $injectedParams = new VideoThumbParams([
+            VideoThumbParams::PARAM_QUALITY_SCALE => 1,
         ]);
 
-        $withParams = (new ThumbParams())->withQualityScale(1);
+        $withParams = (new VideoThumbParams())->withQualityScale(1);
 
         self::assertSame($injectedParams->toArray(), $withParams->toArray());
     }